test(logic): split valid work tiles part1 tests under test/orders
Move getValidWorkOrderTileKeys scenarios from the 743-line root file into
three focused files (basic, build_improvement, prospect), each under the
400-line logic test gate.

Refs #2288
