+++ packages/colonizethis_logic/test/orders/order_suggestion_valid_work_tiles_part1_prospect_test.dart
@@ -1,0 +1,223 @@
+import 'package:colonizethis_test/test.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+void main() {
+  group('getValidWorkOrderTileKeys', () {
+    test(
+      'getValidWorkOrderTileKeysWithVisibility prospect excludes non-mineral '
+      'and already prospected',
+      () {
+        const playerId = 'gp1';
+        const ow = 'oldWorld';
+        const provinceId = '$ow|p1';
+        const grassTile = 'oldWorld|p1|0|0';
+        const ironTile = 'oldWorld|p1|1|0';
+        final player = const Player(
+          id: playerId,
+          displayName: 'GP',
+          isHuman: false,
+        );
+        final tribe = const Tribe(id: 'tribe1', displayName: 'T');
+        final p1 = Province(id: provinceId, regionId: ow, ownerId: 'tribe1');
+        final unit = Unit(
+          id: 'u1',
+          type: kUnitTypeExplorer,
+          ownerId: playerId,
+          locationProvinceId: provinceId,
+          tileKey: grassTile,
+        );
+        final world = WorldState(
+          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
+          oldWorld: RegionData(provinces: [p1], units: [unit]),
+          newWorld: const RegionData(),
+          playerVisibilityByTile: const {
+            playerId: {grassTile: 'fogged', ironTile: 'fogged'},
+          },
+          resourceByTileKey: const {grassTile: 'grain', ironTile: 'iron'},
+          playerProspectedTiles: const {
+            playerId: {ironTile},
+          },
+          tileKeysByRegionAndProvince: {
+            ow: {
+              provinceId: [grassTile, ironTile],
+            },
+          },
+        );
+        final game = Game(
+          id: 'g1',
+          worldState: world,
+          players: [player],
+          tribes: [tribe],
+        );
+        final topology = MapTopology(
+          nodes: const [
+            TopologyNode(
+              id: 'p1',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+          ],
+          edges: const [],
+        );
+        final view = buildPlayerView(game, topology, playerId);
+        final valid = getValidWorkOrderTileKeysWithVisibility(
+          game: game,
+          topology: topology,
+          view: view,
+          unitId: 'u1',
+          workTarget: kWorkTargetProspect,
+          currentOrders: const Orders(),
+        );
+        expect(valid.contains(grassTile), isFalse);
+        expect(valid.contains(ironTile), isFalse);
+      },
+    );
+
+    test(
+      'getValidWorkOrderTileKeysWithVisibility prospect includes eligible tile',
+      () {
+        const playerId = 'gp1';
+        const ow = 'oldWorld';
+        const provinceId = '$ow|p1';
+        const ironTile = 'oldWorld|p1|0|0';
+        final player = const Player(
+          id: playerId,
+          displayName: 'GP',
+          isHuman: false,
+        );
+        final tribe = const Tribe(id: 'tribe1', displayName: 'T');
+        final p1 = Province(id: provinceId, regionId: ow, ownerId: 'tribe1');
+        final unit = Unit(
+          id: 'u1',
+          type: kUnitTypeExplorer,
+          ownerId: playerId,
+          locationProvinceId: provinceId,
+          tileKey: ironTile,
+        );
+        final world = WorldState(
+          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
+          oldWorld: RegionData(provinces: [p1], units: [unit]),
+          newWorld: const RegionData(),
+          playerVisibilityByTile: const {
+            playerId: {ironTile: 'fogged'},
+          },
+          resourceByTileKey: const {ironTile: 'iron'},
+          tileKeysByRegionAndProvince: {
+            ow: {
+              provinceId: [ironTile],
+            },
+          },
+        );
+        final game = Game(
+          id: 'g1',
+          worldState: world,
+          players: [player],
+          tribes: [tribe],
+        );
+        final topology = MapTopology(
+          nodes: const [
+            TopologyNode(
+              id: 'p1',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+          ],
+          edges: const [],
+        );
+        final view = buildPlayerView(game, topology, playerId);
+        final valid = getValidWorkOrderTileKeysWithVisibility(
+          game: game,
+          topology: topology,
+          view: view,
+          unitId: 'u1',
+          workTarget: kWorkTargetProspect,
+          currentOrders: const Orders(),
+        );
+        expect(valid, contains(ironTile));
+      },
+    );
+
+    test(
+      'getValidWorkOrderTileKeysWithVisibility prospect excludes wool on hills '
+      'when tile map marks hills (terrain-only eligibility must not apply)',
+      () {
+        const playerId = 'gp1';
+        const ow = 'oldWorld';
+        const provinceId = '$ow|p1';
+        const woolTile = 'oldWorld|p1|0|0';
+        final player = const Player(
+          id: playerId,
+          displayName: 'GP',
+          isHuman: false,
+        );
+        final tribe = const Tribe(id: 'tribe1', displayName: 'T');
+        final p1 = Province(id: provinceId, regionId: ow, ownerId: 'tribe1');
+        final unit = Unit(
+          id: 'u1',
+          type: kUnitTypeExplorer,
+          ownerId: playerId,
+          locationProvinceId: provinceId,
+          tileKey: woolTile,
+        );
+        final tileMapByRegion = <String, TileMapResult>{
+          ow: TileMapResult(
+            width: 1,
+            height: 1,
+            grid: const [
+              ['p1'],
+            ],
+            terrainGrid: const [
+              [TerrainType.hills],
+            ],
+            resourceGrid: const [
+              [Resource.wool],
+            ],
+          ),
+        };
+        final world = WorldState(
+          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
+          oldWorld: RegionData(provinces: [p1], units: [unit]),
+          newWorld: const RegionData(),
+          playerVisibilityByTile: const {
+            playerId: {woolTile: 'fogged'},
+          },
+          resourceByTileKey: const {woolTile: 'wool'},
+          tileKeysByRegionAndProvince: {
+            ow: {
+              provinceId: [woolTile],
+            },
+          },
+        );
+        final game = Game(
+          id: 'g1',
+          worldState: world,
+          players: [player],
+          tribes: [tribe],
+        );
+        final topology = MapTopology(
+          nodes: const [
+            TopologyNode(
+              id: 'p1',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+          ],
+          edges: const [],
+        );
+        final view = buildPlayerView(game, topology, playerId);
+        final valid = getValidWorkOrderTileKeysWithVisibility(
+          game: game,
+          topology: topology,
+          view: view,
+          unitId: 'u1',
+          workTarget: kWorkTargetProspect,
+          currentOrders: const Orders(),
+          tileMapByRegion: tileMapByRegion,
+        );
+        expect(valid.contains(woolTile), isFalse);
+      },
+    );
+  });
+}
